A few new year edits to resolve common questions
In the past year or so, a couple different CEs have had the same misunderstanding with figure caption guidance, rewriting captions to be consistently fragments or sentences rather than just changing the punctuation. Here's my attempt at a clarification—ultimately directing to you, the PE, since there are so many dependencies.

Per Slack discussion, the direction on compound nouns was too broad and not often enforced anyway (AFAIK). I removed it and added two common culprits to the word list instead: road map and life cycle—we've *been* fine with either (preferring MW) as long as the book is consistent, so why not codify that here.

Feedback welcome!